second commit

+++ src/main/java/eu/fondy/task/dto/ChangeResponseDto.java
@@ -8,17 +8,14 @@ import java.util.Map;
 import java.util.UUID;
 
 
-
 @Builder
 @Data
 public class ChangeResponseDto {
 
-    private long penceSubmitted;
+    private int penceSubmitted;
     private UUID externalID;
     private Map<Integer, Integer> pounds;
     private Map<Integer, Integer> pence;
-//    private String pounds;
-//    private String pence;
     private LocalDateTime dateTime;
 
 }
